+++ src/main/java/qna/common/exception/CannotDeleteException.java
@@ -1,4 +1,4 @@
-package qna;
+package qna.common.exception;
 
 public class CannotDeleteException extends RuntimeException {
     private static final long serialVersionUID = 1L;
