+++ src/main/java/qna/common/exception/ForbiddenException.java
@@ -1,4 +1,4 @@
-package qna;
+package qna.common.exception;
 
 public class ForbiddenException extends RuntimeException {
     public ForbiddenException() {
