+++ src/main/java/qna/common/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package qna;
+package qna.common.exception;
 
 public class NotFoundException extends RuntimeException {
 }
