+++ src/main/java/qna/common/exception/UnAuthenticationException.java
@@ -1,4 +1,4 @@
-package qna;
+package qna.common.exception;
 
 public class UnAuthenticationException extends Exception {
     private static final long serialVersionUID = 1L;
@@ -8,7 +8,7 @@ public class UnAuthenticationException extends Exception {
     }
 
     public UnAuthenticationException(String message, Throwable cause, boolean enableSuppression,
-                                     boolean writableStackTrace) {
+        boolean writableStackTrace) {
         super(message, cause, enableSuppression, writableStackTrace);
     }
