+++ src/main/java/qna/service/QnaService.java
@@ -5,12 +5,10 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import qna.CannotDeleteException;
-import qna.NotFoundException;
+import qna.common.exception.CannotDeleteException;
+import qna.common.exception.NotFoundException;
 import qna.domain.*;
 
-import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
