+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.url=jdbc:h2:~/test;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
 spring.datasource.username=sa
 spring.h2.console.enabled=true
-
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.show-sql=true
+logging.level.org.hibernate.type.descriptor.sql=trace
