+++ src/test/java/qna/domain/AnswerTest.java
@@ -6,8 +6,8 @@ import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import qna.NotFoundException;
-import qna.UnAuthorizedException;
+import qna.common.exception.NotFoundException;
+import qna.common.exception.UnAuthorizedException;
 
 public class AnswerTest {
     public static final Answer A1 = new Answer(UserTest.JAVAJIGI, QuestionTest.Q1, "Answers Contents1");
